+++ interfaces/IBadgePriceOracle.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.8.4;
 
 interface IBadgePriceOracle {
-    function calculateBadgePrice(uint256 level) external view returns (uint256);
+    function calculateBadgePrice(uint8 level) external view returns (uint256);
 }
